third_party/WebKit/LayoutTests/fast/peerconnection/RTCPeerConnection-state.html

<!DOCTYPE html>
<html>
<head>
<title>RTCPeerConnection signalingState</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<script>

async_test(function(test) {
  var pc = new RTCPeerConnection();
  assert_equals(pc.signalingState, 'stable');
  pc.onsignalingstatechange = test.step_func(() => {
    assert_equals(pc.signalingState, 'closed');
    test.done();
  });
  pc.close();
}, 'Tests the RTCPeerConnection "stable" and "closed" signalingState strings.');

</script>
</body>
</html>
